requirements.txt
streamlit==1.28.0
openai==0.28.0
PyPDF2==3.0.1
Pillow==10.0.1